fix: change op type to float64 in fake gateway client

+++ internal/pkg/flink/test/generators/results.go
@@ -128,7 +128,7 @@ func MockResultRow(columnDetails []flinkgatewayv1alpha1.ColumnDetails) *rapid.Ge
 			items = append(items, GetResultItemGeneratorForType(column.GetType()).Draw(t, "a field"))
 		}
 		return map[string]any{
-			"op":  rapid.Float64Range(0, 3).Draw(t, "an operation"),
+			"op":  float64(rapid.IntRange(0, 3).Draw(t, "an operation")),
 			"row": items,
 		}
 	})
